added register api and button
